Refactor: 인터페이스 의존 적용
